+++ src/array/empty/method.ts
@@ -1,3 +1,5 @@
-const method = (arr: any[]) => arr.length === 0;
+const method = (arr: any[]) => {
+  arr.length = 0;
+};
 
 export = method;