+++ src/array/isEmpty/method.ts
@@ -1,0 +1,3 @@
+const method = (arr: any[]) => arr.length === 0;
+
+export = method;