+++ src/array/methods.ts
@@ -23,6 +23,7 @@ import * as implode from "./implode/method";
 import * as indexOfAll from "./indexOfAll/method";
 import * as initial from "./initial/method";
 import * as intersect from "./intersect/method";
+import * as isEmpty from "./isEmpty/method";
 import * as last from "./last/method";
 import * as limit from "./limit/method";
 import * as max from "./max/method";
@@ -81,6 +82,7 @@ export {
   indexOfAll,
   initial,
   intersect,
+  isEmpty,
   last,
   limit,
   max,