+++ src/object/empty/method.ts
@@ -1,5 +1,10 @@
 import * as size from "../size/method";
 
-const method = (obj: object) => size(obj) === 0;
+const method = (obj: object) => {
+  const keys = Object.keys(obj);
+
+  for (let i = 0; i < keys.length; i++)
+    delete (obj as any)[keys[i]];
+};
 
 export = method;